MVP: Fetching Purchased Steam Games by Account
As of this commit, Steam doesn't offer this datapoint in their API, so I found a workaround: Javascript fires an AJAX request to a local PHP script, which in turn fires a request to Steam for an XML list of games by account. PHP script prints the XML, and the Javascript parses and reformats it into an array of objects.